refact: remove account data from nostr api logic
